requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
numpy==1.26.2
pandas==2.1.4
pillow==9.4.0
tqdm==4.66.1